ui/gtk: move guest mouse cursor after host cursor hits the edge of the display
When the host mouse cursor hits the boundary of the surface, the guest
mouse cursor also stops moving to one direction even if it is still in
the middle of the guest display. This is because dX and/or dY won't be
generated anymore. So it is needed to create those delta position data
artificially when the mouse cursor is still in move.

Cc: Marc-André Lureau <marcandre.lureau@redhat.com>
Cc: Vivek Kasireddy <vivek.kasireddy@intel.com>
Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>