virtio-gpu-udmabuf: Create dmabuf from mr associated with a VFIO dev
Once we have the ramblock and the associated mr, we try to find the
VFIO device that owns the mr and also determine the relevant VFIO
region and index. We then provide this info (i.e, region index)
along with offsets and lengths to the VFIO PCI driver (via
VFIO_DEVICE_FEATURE_GET | VFIO_DEVICE_FEATURE_DMA_BUF) to get
a new dmabuf fd created.

Signed-off-by: Vivek Kasireddy <vivek.kasireddy@intel.com>