ui/gtk: Page number of '-1' is not a valid page number
Page number of '-1' means the page does not belong to the
notebook so it shouldn't be used as a valid page number in
gd_vc_find_by_page. It should just return null if that is
the case.

This change, however, will cause a segfault during detaching
/untabifying process in gtk_release_modifiers because the
current VC's page number suddenly becomes '-1' as soon as
it is detached then this invalid page number will eventually
bring in a null VC. So gtk_relaese_modifiers should do the
null check on the VC and just returns if VC == null.

Cc: Gerd Hoffmann <kraxel@redhat.com>
Cc: Marc-André Lureau <marcandre.lureau@redhat.com>
Signed-off-by: Vivek Kasireddy <vivek.kasireddy@intel.com>
Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>