ui/gtk: Fix lookup and segfault for invalid page numbers
A page number of -1 is invalid and can refer to multiple
unattached Virtual Consoles (VCs). This patch ensures negative
page numbers are no longer treated as valid indices.

To prevent a regression, a null check is added to
gtk_release_modifiers. This avoids a segmentation fault during
the "untabify" process, where a VC's page number is temporarily
set to -1 immediately after detaching.

Cc: Gerd Hoffmann <kraxel@redhat.com>
Cc: Marc-André Lureau <marcandre.lureau@redhat.com>
Cc: Vivek Kasireddy <vivek.kasireddy@intel.com>
Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>