d/microvm-devices.mak: disable VIRTIO_MEM and VIRTIO_PMEM devices
these devices has been enabled, apparently wronly, in a response
to #992029, even if they weren't actually requested.

However, these devices both require PCI which is not enabled, and now
the binary fails to link since PCI-related functionality isn't being
compiled in anymore in case it is not enabled, after change in 9.1
including "hw-virtio-move-stubs-out-of-stubs.patch" which we're
reverting in order to make it build again.

If someone needs virtio_mem or virtio_pmem with microvm, we'll have
to re-check this.