ui/gtk: Added an input mode
In a multi-seat scenario where multiple keyboards and mice are
connected to the host but some are dedicated for the guests only
(through pass through mode) and some are only for the host, there
is a strong use case where a customer does not want a HID device
connected to the host to be able to control the guest.

In such a scenario, neither should we bind any input events to
Qemu UI, nor should we show menu options like "Grab on Hover" or
"Grab Input". This patch adds a GTK command line option called
"input". It can be set like this:

  gtk,input=off/on

If set to on or completely left out, it will default to normal
operation where host HID devices can control the guests. However,
if turned off, then host HID devices will not be able to control
the guest windows.

Cc: Dongwon Kim <dongwon.kim@intel.com>
Cc: Vivek Kasireddy <vivek.kasireddy@intel.com>
Cc: Gerd Hoffmann <kraxel@redhat.com>
Cc: Marc-André Lureau <marcandre.lureau@redhat.com>>
Cc: Daniel P. Berrangé <berrange@redhat.com>
Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>