virtio-gpu: Replace the surface with null surface only when resetting
Update the current surface with a place holder surface only when
virtio-gpu device is being reset. It is required to keep the most
recent frame on the display even if the resource associated with it
is unreferenced. Otherwise, frequent blnking will happen whenever
the resource associated with the scanout is unreferenced.

Cc: Marc-André Lureau <marcandre.lureau@redhat.com>
Cc: Vivek Kasireddy <vivek.kasireddy@intel.com>
Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>