Carry Ubuntu specific machine types
Since Ubuntu is a downstream of qemu carrying patches it needs custom machine
types to be able to identify and manage the delta that might affect machine
types.

This is an important piece to keep cross release migration supported for any
downstream.

Since the p->t transition these types are mostly stable copies of the upstream
type (in the past this was more unstable upstream, so there was more delta),
but they need to stay specific to reflect the delta we have. And even more so
to have something to base off for affecting SRU changes.

Also add a hint if instantiating fails due to now unsupported old guest
types (LP: #1637936).

Package maintainers please see https://wiki.ubuntu.com/QemuKVMMigration when
maintaining this patch on SRU, merge or other packaging activity.
While support on a type is dropped with the Release going EOL we never drop the
type itself as long as it is maintainable. This will give people an extra
chance to migrate and avoid issues like LP: 1802944.

##

This later on got extended by further ubuntu specific machine type changes:
LP 1776189: Add a -hpb Ubuntu specific machine type suffix

This works already fine on commandline, but Libvirt and other stacks above
have no exploitation yet. Using a machine type has the benefit of being already
controllable by most upper layer software like Libvirt (type= in os tag) but
even up to Openstack (nova.conf or per image metadata on hw_machine_type).

This is based on a discussion:
   https://bugs.launchpad.net/ubuntu/+source/qemu/+bug/1769053

A similar change is in CentOS/RH (there the default is switched, without
even a way to go back.
But since this can cause issues e.g. when migrating
across hosts with different characteristics, it is not set as the default
in Ubuntu with this change.

Further we want to avoid "machine type proliferation", so we certainly won't
add a type for every feature. But using a huge guest is more common and
otherwise not yet achievable.

From Noble onwards, we do not add the -hpb machine variants
because they are not needed by OpenStack anymore.  For more information, see:

https://bugs.launchpad.net/ubuntu/+source/qemu/+bug/1769053
https://bugs.launchpad.net/ubuntu/+source/qemu/+bug/2045592

##

This later on got extended by further ubuntu specific machine type changes:
LP 1761372: special type for ppc64 meltdown/spectre defaults

Upstream 2.12 is not yet set in stone (almost but not full), and we ship 2.11
with backports. So we don't want to make a 2.12 machine type fully recommended
yet.
PPC was following x86 in providing a non default convenience type that has the
spectre/meltdown flags toggled - in bug 1761372 we were requested to carry the
same - but we agreed to do so as a 2.11 based type.

Note I: x86 changes CPU types with -IBRS suffix, power chose to change machine
types.

Note II: this change can be squashed into ubuntu-machine-types.patch >=2.12
where the base content will exist in the upstream source instead of
patches on top.

##

[1] introduced a major regression into the 4.0 types by setting split
irqchip to be the default. This was corrected by [2] and the fix further
modified by [3] which overall adds a 4.0.1 machine type in qemu 4.1 (not
yet released) and probably eventually stable branches.
We will follow upstream with the upstream types, but the Ubuntu types so
far didn't release a 4.0 type yet so for us we can fix it on the initial
release right away.

[1]: https://git.qemu.org/?p=qemu.git;a=commit;h=b2fc91db
[2]: https://git.qemu.org/?p=qemu.git;a=commit;h=c87759ce
[3]: https://git.qemu.org/?p=qemu.git;a=commit;h=8e8cbed0

##

In Jammy (22.04) the former logic to warn about dropped types (P/U/V) no
more applied cleanly, but since on one hand we warned about those now very
outdated types long enough and we no more drop types to retain max
compatibility this snippet was intentionally dropped. Users trying to use
e.g. pc-i440fx-precise will still get a warning that the type does not
exist and a hint how to check which are supported. Just no extra warning
that Precise/Utopic/Vivid are out of support.

##

For Jammy and Mantic, qemu-system-amd64 needed to be patched to
support more than 288 vCPUs inside a VM.  This was done by backporting
a specific upstream change.  On top of that, and because we offer our
own machine types to Ubuntu users, we also needed to create special
machine types for the existing ones on Jammy and Mantic (that had
already been released).  This is why you will see the "-maxcpus"
variant below.

For more context on the changes that were made, see:

https://bugs.launchpad.net/ubuntu/+source/qemu/+bug/2012763

##

Add v2 ubuntu machine types for arch caps handling

Across ubuntu releases before 25.10 questing we have to fix
(LP: #2131822) the handling of arch-caps. But doing so would
break existing guest, therefore this new behavior is gated
by newly created -v2 ubuntu machine types. These types and
handling needs to be known to subsequent releases to ensure
they can continue to host these guests e.g. in migrations

Bug-Ubuntu: https://bugs.launchpad.net/bugs/2131822

##

At the release of Ubuntu Questing, we mistakenly based I440FX
Ubuntu Questing machine type on v10.0 istead of the v10.1. To
fix this issue, we add a new Questing I440FX v2 machine
to be based on v10.1 instead of v10.0.

##

Original-Author: Serge Hallyn <serge.hallyn@ubuntu.com>
Original-Bug-Ubuntu: https://bugs.launchpad.net/ubuntu/+source/qemu/+bug/1304107
Author: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Author: Sergio Durigan Junior <sergio.durigan@canonical.com>
Bug-Ubuntu: https://bugs.launchpad.net/bugs/1776189
Bug-Ubuntu: https://bugs.launchpad.net/bugs/1761372
Bug-Ubuntu: https://bugs.launchpad.net/bugs/1769053
Bug-Ubuntu: https://bugs.launchpad.net/bugs/2012763
Bug-Ubuntu: https://bugs.launchpad.net/bugs/2131822
Forwarded: not-needed
Forward-info: downstream decision