ui/gtk-egl: Skipping frame drawing if not necessary
A flag 'new_cursor' would never be cleared in some corner case
where the cursor is not inside the frame or is not moved in
the next refresh cycle. This situation forces it to keep drawing
the cursor-less frame at every refresh cycle. This is a buggy
situation so needs to be adjusted as followed.

1. Draw the cursor then blend it with the existing frame if
   the cursor is moved AND the new location is inside the
   frame

2. Draw only the existing frame without cursor only if
   the cursor is moved AND the cursor was drawn during the
   previous cycle BUT the new location is not within the
   frame boundary.

3. Skipping redrawing anything otherwise.

Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>