aarch64-linux-user: turn off pauth by default
By default, qemu emulates pauth feature (pauth_impdef) on aarch64 CPU.
However, it costs quite a lot of host CPU cycles to implement, which
results in significant slowdowns of the guest code which uses pauth.
Since more and more code uses uses pauth (which is implemented in-cpu
on aarch64, and costs basically nothing on real hardware), we observe
more and more slowness, to the effect of being impractical.

Turn off pauth emulation by default at least for the linux-user case,
so that foreign-arch aarch64 binaries aren't dramatically slow.  For
the qemu system mode, things has to be backwards-compatible and
migratable between qemu versions, so things aren't that simple.