u-boot-sam460ex: fix "failed" puts
There's a global char *failed="failed" variable,
which is used below the code in question like
 puts(failed).
However, the code in question, #ifdef'ed, uses
the variable of the same name but with different
type (u8).  Which results in a failed compilation.
Rename the local variable `failed' to `ret'.

(Actually, there's no need to store the result of
the function call in a variable, since it is only
used once anyway).

Signed-off-by: Michael Tokarev <mjt@tls.msk.ru>