u-boot-sam460ex: remove compiler-gccN.h #include
include/linux/compiler*.h are taken from old linux, and
only supports gcc4.  However, the gcc-version-dependent
definitions in there aren't actually used in the code,
except of uninitialized_var macro - which is defined in
the fallback case to the same value.  So there's no
point in including the gcc-version-specific header,
which fails when building this code with a more recent
version of gcc.

This patch removes just the #include statement
(and the helper definitions), but not compiler-gcc4.h
header itself.  This makes the change small, and allows
compilation of u-boot-sam460ex code by something more
recent than ancient gcc.

Signed-off-by: Michael Tokarev <mjt@tls.msk.ru>